+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 1.1.0
 * Added support for a nested activator in `PerControllerConfigActivator`
+* Added `StringCollectionConstraint` and `EnumConstraint` which can be used as custom route constraints.
 
 ### 1.0 - 23 March 2015
 * First release.